+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- Repository -> Init... menuitem, which creates a new repository.
 - Progress bar for slow git operations.
 - GG now respects the `snapshot.auto-update-stale` setting. Additionally, when first opening a repo, it will always update the working copy if it's stale.